fix: Prevent destruction of topic ACLs [DEVOP-5504]
Currently there are a lot of ACLs that share the same ID due
to them using the `PREFIX` pattern type, but still created in
state for all the topics utilising that `PREFIX` which can lead
to accidental unavailability if you delete one of the topics that
utilises that ACL identifier.

In order to block accidental deletion and unavailability, we will
set the `prevent_destroy` lifecycle policy so that any destruction
events will rather error out the plan/apply.

The long-term solution will be of two parts;
1. Migrate the ACL resources to their own module, such that in cases
   like `PREFIX` patterns can be managed once, instead of `n` times,
2. A thorough design session to implement tighter access controls in
   order to move towards a better segmented access policy.

Refs: #DEVOP-5504

Signed-off-by: Christian Witts <christian@honestbank.com>