fix: Confluent CLI v4 has some breaking changes
As we don't version pin the CLI being installed, and we use
`latest`, there are breaking changes moving from the _v3.x_
CLI that was last run ~18 months ago. One of which is that the
the [schema registry](#1) no longer needs to be created
separately as it always-enabled in Confluent Cloud environments.

Due to that breaking change, we will remove the `local-exec`
provisioners from attempting to run the commands required of the prior
_v3.x_ CLI version so that we no longer get failures while trying
to execute (sub)commands that no longer exist.

[1] https://docs.confluent.io/confluent-cli/current/migrate.html#cli-breaking-changes-v4

Refs: #DEVOP-5504

Signed-off-by: Christian Witts <christian@honestbank.com>